networking loads/initializes game scene
This means that I can move away from UI/Menu stuff and now work on
placing GridObjects.

+++ Assets/Scripts/Game/Grid/GridObject.cs
@@ -10,7 +10,7 @@ namespace BattlePark {
 			public int X;
 			public int Y;
 			public int Z;
-			public int Owner;
+			public long Owner;
 		}
 	
 		public Direction Direction;
@@ -21,7 +21,7 @@ namespace BattlePark {
 		public Vector3[] OccupiedOffsets = { Vector3.zero };
 		#endregion
 	
-		public int Owner;
+		public long Owner;
 	
 		public Grid Grid;
